requirements.txt
Django==5.0.6
environs==11.0.0
pillow==10.3.0
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
dj-config-url==0.1.1
